Add a Linux system and a IGpioPin implementation for it.
This is a wrapper for the standard Linux GPIO interface via sysfs, which
should work on any platform whose driver provides the sysfs interface.
This includes the Raspberry Pi, though that chipset has a faster memory
mapped IO interface which we may want to add special support for later
in a separate "BCM2708" system.